remove general regulations from regulation group library generated from active plan
This way there shouldn't be any chance a plan regulation group widget to be created from a general regulation,
preventing the case where the plan is counted as a plan feature linked to a regulation group. This change is also
a fix in itself because it shouldn't be possible to assign general regulations for plan features.