optimize plugin by resaving features only when some of their attribute has changed
Resaving is skipped by checking the 'modified' attribute of a model. Modified is True by default for new models and False by default
when initializing a model from DB. When a widget or form is saved as a model in the UI, it is checked whether model fields that
have a corresponding attribute in the linked layer have changed, and based on the result value of 'modified' is set.