fix plan objects dock crashing when changing plan matter
When changing plan matter, a filter subsetString gets set to plan feature layers
resulting in invalid features when they are about to be accessed in plan features
dock. This fix makes sure a feature is valid before accessing its attributes when
reacting to feature selection changes in the plan features dock.

When a new active plan is set in PlanManager, the plan features dock gets rebuilt
which should ensure correctness of the selected_plan_feature_ids set.